fix: Make package includes work
When using require on vue components instead of import, the component
may not always be under the expected name. In this case, for Slideout
it's 'default', and for videoPlayer it's 'videoPlayer'. Since all we
need is the component for our list, we can just add the name and use
that. This means we can get rid of our file imports for now.