chore: Collapse HapticVideoPlayer into main app
Syncydink IS a haptic video player. There's no reason to separate the
components at the level we had them at. This was making for a ton of
work passing haptics commands between buttplug, the upcoming encoder
work, and wherever else we may want to take them. This should allow us
to continue avoiding Vuex for a while, since our main time and haptics
generator is now at the top level of the application.